+++ src/BDSOutput.cc
@@ -647,18 +647,14 @@ void BDSOutput::FillSamplerHits(const BDSHitsCollectionSampler* hits,
     {
       const BDSHitSampler* hit = (*hits)[i];
       G4int samplerID = hit->samplerID;
-      samplerID += 1; // offset index by one due to primary branch.
       samplerTrees[samplerID]->Fill(hit, storeSamplerMass, storeSamplerCharge, storeSamplerPolarCoords, storeSamplerIon, storeSamplerRigidity, storeSamplerKineticEnergy);
     }
 
   // extra information
-  G4bool firstSampler = true;
-  for (auto& sampler : samplerTrees)
+  if (storeSamplerIon)
     {
-      if (firstSampler) // skip primaries (1st sampler) as it always has extras filled in
-	{firstSampler = false; continue;}
-      if (storeSamplerIon)
-        {sampler->FillIon();}
+      for (auto& sampler : samplerTrees)
+	{sampler->FillIon();}
     }
 }